More leak fixes, largely in spatial. Spatial extension objects weren't copying their pointers when copied, and weren't deleting them when done.

+++ src/sbml/packages/spatial/sbml/SampledField.h
@@ -459,7 +459,7 @@ public:
    * @li LIBSBML_OPERATION_SUCCESS
    * @li LIBSBML_INVALID_ATTRIBUTE_VALUE
    */
-  virtual int setSamples(int* inArray, int arrayLength);
+  virtual int setSamples(int* inArray, size_t arrayLength);
 
 
   /**
